ACQ-751: moving out delivery-option-messages from components to utils.
