+++ jest.config.js
@@ -1,9 +1,12 @@
 module.exports = {
-	testPathIgnorePatterns: ['/node_modules/', '/bower_components/', '/cypress/'],
+	testPathIgnorePatterns: ['/node_modules/', '/bower_components/', '/cypress/', '/test/'],
 	transform: {
 		'.(js|jsx)': '@sucrase/jest-plugin',
 	},
-	testMatch: ['**/components/**/?(*.)+(spec|test).[tj]s?(x)'],
+	testMatch: [
+		'**/components/**/?(*.)+(spec|test).[tj]s?(x)',
+		'**/utils/**/?(*.)+(spec|test).[tj]s?(x)',
+	],
 	snapshotSerializers: ['jest-serializer-html'],
 	setupFilesAfterEnv: ['<rootDir>/test-jest/helpers/setup.js'],
 };
